requirements.txt
# Cloud Backbone Requirements
pyyaml>=6.0
requests>=2.31.0
boto3>=1.28.0  # AWS SDK for cost optimization
psutil>=5.9.0  # System monitoring
schedule>=1.2.0  # Task scheduling
